+++ src/index.js
@@ -2,11 +2,19 @@
 export * as utils from './utils'
 export * from './utils'
 
+// hooks
+export * as hooks from './hooks'
+export * from './hooks'
+
+// theme
+export * as theme from './theme'
+export * from './theme'
+
 // components
 export { default as Button } from './Button'
 export { default as Checkbox } from './Checkbox'
 export { default as Icon } from './Icon'
 export { default as MenuItem } from './MenuItem'
-export { default as Message } from './Message'
+export { default as message } from './message'
 export { default as Select } from './Select'
 export { default as TextField } from './TextField'